UFC Weight Class
This assembly code prompts the user to enter their weight in pounds and stores the input value in the weight variable. The program then compares the user's weight to predetermined weight ranges for each of the UFC weight classes, starting with the lowest class (Flyweight) and progressing to the highest class (Heavyweight).

When the user's weight falls within a certain range, the program will display the name of the corresponding weight class to the user using the WriteString procedure. Once the weight class is determined, the program will display a message to the user indicating which weight class they belong to.

Overall, this code allows the user to input their weight and receive information about their corresponding UFC weight class.